requirements.txt
customtkinter
requests
python-dotenv
srt
